src: M1 -pi rollout batch 1 (atk/value, adew, apt/{apt,suite,tree}, controllers); .ch drift + rock fixes; suite/treev attribute pairs expanded at dispatch sites
FossilOrigin-Name: d111e78211b104391cc17b38ef60a33ab8ae8d568817da7af84ba4486d361038

+++ src/atk/adew/lset.c
@@ -68,7 +68,7 @@ struct lset *self;
 	printf("Posting %d for %s (%s)\n",(long)self->dobj,buf,class_GetTypeName(self->dobj)); 
 #endif /* DEBUG */
 	rm_PostResource(al,(long)self->dobj,a_vp);
-	lset_Put(self,a_name,a_atomlist,al);
+	lset_Put(self,a_name,a_atomlist,(long)al);
     }
 }
 static struct dataobject *getregisteredobject(self)
